multithread_multiprocess/pyproject.toml
[project]
name = "multithread-multiprocess"
version = "0.1.0"
description = ""
authors = [
    {name = "Jorge Alves",email = "jorgeandrealves@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow (>=11.1.0,<12.0.0)",
    "pandas (>=2.2.3,<3.0.0)"
]

[tool.poetry]
packages = [{include = "multithread_multiprocess", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
